Small clarity and typo edits. Confirmed working
This adds a small "Next steps" section on the getting started to suggest people
visit the other datasets and a message about "You can't change the server data, so explore freely"
at the bottom.

Fixed some very small typos and a missing sentance in the other ones.

I can confirm that all of these examples are working with the current master head of
QCFractal at commit 9b4d1e1